Ajuste no nome do pacote para utilizar um sufixo que define que ẽ um projeto open source
