--- src/main/java/com/ueby/loteria/crawlers/CaixaCrawlerStub.java
@@ -1,19 +1,0 @@
-package com.ueby.loteria.crawlers;
-
-import lombok.*;
-
-/**
- * @author Algarves, Algarves
- */
-@Getter
-@Setter
-@NoArgsConstructor(access = AccessLevel.PUBLIC)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder(builderClassName = "Builder", setterPrefix = "with")
-public class CaixaCrawlerStub {
-
-  private Integer concurso;
-
-  private String result;
-
-}
