+++ src/main/java/com/ueby/oss/loteria/crawlers/CaixaGameType.java
@@ -1,0 +1,28 @@
+package com.ueby.oss.loteria.crawlers;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * @author Algarves, Khristian
+ */
+@Getter
+@AllArgsConstructor
+public enum CaixaGameType {
+  
+  // @formatter:off
+  MEGA_SENA("/megasena/"),
+  QUINA("/quina/"),
+  LOTOFACIL("/lotofacil/"),
+  LOTOMANIA("/lotomania/"),
+  DUPLA_SENA("/duplasena/"),
+  TIMEMANIA("/timemania/"),
+  DIA_DE_SORTE("/diadesorte/"),
+  LOTECA("/loteca/"),
+  LOTERIA_FEDERAL("/federal/"),
+  SUPER_SETE("/supersete/");
+  // @formatter:on
+
+  final String path;
+
+}
