+++ src/main/java/com/ueby/oss/loteria/crawlers/exception/CaixaWebCrawlerException.java
@@ -1,0 +1,18 @@
+package com.ueby.oss.loteria.crawlers.exception;
+
+/**
+ * @author Algarves, Khristian
+ */
+public class CaixaWebCrawlerException extends Exception {
+
+  static final long serialVersionUID = 1L;
+
+  public CaixaWebCrawlerException(String message) {
+    super(message);
+  }
+
+  public CaixaWebCrawlerException(String message, Throwable throwable) {
+    super(message, throwable);
+  }
+
+}
