--- src/test/java/com/ueby/loteria/crawlers/game/LotofacilWebCrawlerTest.java
@@ -1,47 +1,0 @@
-package com.ueby.loteria.crawlers.game;
-
-import com.ueby.loteria.crawlers.CaixaGameType;
-import com.ueby.loteria.crawlers.IntegrationTest;
-import com.ueby.loteria.crawlers.service.CaixaWebCrawlerService;
-import org.junit.jupiter.api.*;
-
-/**
- * @author Algarves, Khristian
- */
-@IntegrationTest
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class LotofacilWebCrawlerTest {
-
-  private CaixaWebCrawlerService crawler;
-  private final CaixaGameType caixaGameType = CaixaGameType.LOTOFACIL;
-
-  @BeforeAll
-  void setUp() {
-    crawler = new CaixaWebCrawlerService(caixaGameType);
-  }
-
-  @DisplayName("Lotofacil - Request by last contest")
-  @Test
-  void shouldLotofacilWebCrawlerByLastContest() {
-    crawler.byLastContest();
-
-    CaixaGameResult caixaGameResult = crawler.getMatchResult();
-
-    Assertions.assertNotNull(caixaGameResult.getNumero());
-    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
-  }
-
-  @DisplayName("Lotofacil - Request by number contest")
-  @Test
-  void shouldLotofacilWebCrawlerByNumberContest() {
-    final Integer contest = 1000;
-
-    crawler.byContestNumber(contest);
-
-    CaixaGameResult caixaGameResult = crawler.getMatchResult();
-
-    Assertions.assertEquals(contest, caixaGameResult.getNumero());
-    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
-  }
-
-}
