--- src/test/java/com/ueby/loteria/crawlers/game/QuinaWebCrawlerTest.java
@@ -1,47 +1,0 @@
-package com.ueby.loteria.crawlers.game;
-
-import com.ueby.loteria.crawlers.CaixaGameType;
-import com.ueby.loteria.crawlers.IntegrationTest;
-import com.ueby.loteria.crawlers.service.CaixaWebCrawlerService;
-import org.junit.jupiter.api.*;
-
-/**
- * @author Algarves, Khristian
- */
-@IntegrationTest
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class QuinaWebCrawlerTest {
-
-  private CaixaWebCrawlerService crawler;
-  private final CaixaGameType caixaGameType = CaixaGameType.QUINA;
-
-  @BeforeAll
-  void setUp() {
-    crawler = new CaixaWebCrawlerService(caixaGameType);
-  }
-
-  @DisplayName("Quina - Request by last contest")
-  @Test
-  void shouldQuinaWebCrawlerByLastContest() {
-    crawler.byLastContest();
-
-    CaixaGameResult caixaGameResult = crawler.getMatchResult();
-
-    Assertions.assertNotNull(caixaGameResult.getNumero());
-    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
-  }
-
-  @DisplayName("Quina - Request by contest number")
-  @Test
-  void shouldQuinaWebCrawlerByContestNumber() {
-    final Integer contest = 2110;
-
-    crawler.byContestNumber(contest);
-
-    CaixaGameResult caixaGameResult = crawler.getMatchResult();
-
-    Assertions.assertEquals(contest, caixaGameResult.getNumero());
-    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
-  }
-
-}
