--- src/test/java/com/ueby/loteria/crawlers/util/JsonObjectDto.java
@@ -1,15 +1,0 @@
-package com.ueby.loteria.crawlers.util;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-/**
- * @author Algarves, Khristian
- */
-@NoArgsConstructor
-class JsonObjectDto {
-
-  @Getter
-  private String key;
-
-}
