+++ src/test/java/com/ueby/oss/loteria/crawlers/IntegrationTest.java
@@ -1,0 +1,14 @@
+package com.ueby.oss.loteria.crawlers;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.junit.jupiter.api.Tag;
+
+@Tag("integrationTest")
+@Target({ElementType.TYPE, ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface IntegrationTest {
+
+}
