+++ src/test/java/com/ueby/oss/loteria/crawlers/game/DuplasenaWebCrawlerTest.java
@@ -1,0 +1,47 @@
+package com.ueby.oss.loteria.crawlers.game;
+
+import com.ueby.oss.loteria.crawlers.CaixaGameType;
+import com.ueby.oss.loteria.crawlers.IntegrationTest;
+import com.ueby.oss.loteria.crawlers.service.CaixaWebCrawlerService;
+import org.junit.jupiter.api.*;
+
+/**
+ * @author Algarves, Khristian
+ */
+@IntegrationTest
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class DuplasenaWebCrawlerTest {
+
+  private CaixaWebCrawlerService crawler;
+  private final CaixaGameType caixaGameType = CaixaGameType.DUPLA_SENA;
+
+  @BeforeAll
+  void setUp() {
+    crawler = new CaixaWebCrawlerService(caixaGameType);
+  }
+
+  @DisplayName("Duplasena - Request by last contest")
+  @Test
+  void shouldDuplasenaWebCrawlerByLastContest() {
+    crawler.byLastContest();
+
+    CaixaGameResult caixaGameResult = crawler.getMatchResult();
+
+    Assertions.assertNotNull(caixaGameResult.getNumero());
+    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
+  }
+
+  @DisplayName("Duplasena - Request by contest number")
+  @Test
+  void shouldDuplasenaWebCrawlerByContestNumber() {
+    final Integer contest = 1709;
+
+    crawler.byContestNumber(contest);
+
+    CaixaGameResult caixaGameResult = crawler.getMatchResult();
+
+    Assertions.assertEquals(contest, caixaGameResult.getNumero());
+    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
+  }
+
+}
