+++ src/test/java/com/ueby/oss/loteria/crawlers/game/LotecaWebCrawlerTest.java
@@ -1,0 +1,47 @@
+package com.ueby.oss.loteria.crawlers.game;
+
+import com.ueby.oss.loteria.crawlers.CaixaGameType;
+import com.ueby.oss.loteria.crawlers.IntegrationTest;
+import com.ueby.oss.loteria.crawlers.service.CaixaWebCrawlerService;
+import org.junit.jupiter.api.*;
+
+/**
+ * @author Algarves, Khristian
+ */
+@IntegrationTest
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class LotecaWebCrawlerTest {
+
+  private CaixaWebCrawlerService crawler;
+  private final CaixaGameType caixaGameType = CaixaGameType.LOTECA;
+
+  @BeforeAll
+  void setUp() {
+    crawler = new CaixaWebCrawlerService(caixaGameType);
+  }
+
+  @DisplayName("Loteca - Request by last contest")
+  @Test
+  void shouldLotecaWebCrawlerByLastContest() {
+    crawler.byLastContest();
+
+    CaixaGameResult caixaGameResult = crawler.getMatchResult();
+
+    Assertions.assertNotNull(caixaGameResult.getNumero());
+    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
+  }
+
+  @DisplayName("Loteca - Request by contest number")
+  @Test
+  void shouldLotecaWebCrawlerByContestNumber() {
+    final Integer contest = 850;
+
+    crawler.byContestNumber(contest);
+
+    CaixaGameResult caixaGameResult = crawler.getMatchResult();
+
+    Assertions.assertEquals(contest, caixaGameResult.getNumero());
+    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
+  }
+
+}
