+++ src/test/java/com/ueby/oss/loteria/crawlers/game/MegasenaWebCrawlerTest.java
@@ -1,0 +1,47 @@
+package com.ueby.oss.loteria.crawlers.game;
+
+import com.ueby.oss.loteria.crawlers.CaixaGameType;
+import com.ueby.oss.loteria.crawlers.IntegrationTest;
+import com.ueby.oss.loteria.crawlers.service.CaixaWebCrawlerService;
+import org.junit.jupiter.api.*;
+
+/**
+ * @author Algarves, Khristian
+ */
+@IntegrationTest
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class MegasenaWebCrawlerTest {
+
+  private CaixaWebCrawlerService crawler;
+  private final CaixaGameType caixaGameType = CaixaGameType.MEGA_SENA;
+
+  @BeforeAll
+  void setUp() {
+    crawler = new CaixaWebCrawlerService(caixaGameType);
+  }
+
+  @DisplayName("Megasena - Request by last contest")
+  @Test
+  void shouldMegasenaWebCrawlerByLastContest() {
+    crawler.byLastContest();
+
+    CaixaGameResult caixaGameResult = crawler.getMatchResult();
+
+    Assertions.assertNotNull(caixaGameResult.getNumero());
+    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
+  }
+
+  @DisplayName("Megasena - Request by contest number")
+  @Test
+  void shouldMegasenaWebCrawlerByContestNumber() {
+    final Integer contest = 2110;
+
+    crawler.byContestNumber(contest);
+
+    CaixaGameResult caixaGameResult = crawler.getMatchResult();
+
+    Assertions.assertEquals(contest, caixaGameResult.getNumero());
+    Assertions.assertEquals(caixaGameType, caixaGameResult.getTipoJogo());
+  }
+
+}
