+++ src/test/java/com/ueby/oss/loteria/crawlers/util/JsonObjectDto.java
@@ -1,0 +1,15 @@
+package com.ueby.oss.loteria.crawlers.util;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+/**
+ * @author Algarves, Khristian
+ */
+@NoArgsConstructor
+class JsonObjectDto {
+
+  @Getter
+  private String key;
+
+}
